requirements.txt
discord.py
python.env
